requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
azure-core==1.30.0
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
isodate==0.6.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.12.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
torch==2.2.0+cpu
torchvision==0.17.0+cpu
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.27.0.post1
gunicorn